Revert "Diff update redesign tracking"
