+++ core/src/main/java/com/orientechnologies/orient/core/db/record/ORecordOperation.java
@@ -45,8 +45,6 @@ public class ORecordOperation implements Comparable {
   public ORecordCallback<Long>    createdCallback = null;
   public ORecordCallback<Integer> updatedCallback = null;
 
-  private Object resultData;
-
   public ORecordOperation() {
   }
 
@@ -82,10 +80,6 @@ public class ORecordOperation implements Comparable {
 
   public ORecord getRecord() {
     return record != null ? record.getRecord() : null;
-  }
-
-  public OIdentifiable getRecordContainer() {
-    return record;
   }
 
   public ORID getRID() {
@@ -133,13 +127,4 @@ public class ORecordOperation implements Comparable {
   public int compareTo(Object o) {
     return record.compareTo(((ORecordOperation) o).record);
   }
-
-  public Object getResultData() {
-    return resultData;
-  }
-
-  public void setResultData(Object resultData) {
-    this.resultData = resultData;
-  }
-
 }
