+++ core/src/main/java/com/orientechnologies/orient/core/record/impl/ODocumentEntry.java
@@ -19,19 +19,13 @@
  */
 package com.orientechnologies.orient.core.record.impl;
 
-import com.orientechnologies.orient.core.db.record.OMultiValueChangeEvent;
 import com.orientechnologies.orient.core.db.record.OMultiValueChangeTimeLine;
 import com.orientechnologies.orient.core.metadata.schema.OProperty;
 import com.orientechnologies.orient.core.metadata.schema.OType;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-
 /**
  * Document entry. Used by ODocument.
- *
+ * 
  * @author Emanuele Tagliaferri
  * @since 2.1
  */
@@ -53,115 +47,6 @@ public class ODocumentEntry {
 
   public boolean isChanged() {
     return changed;
-  }
-
-  public boolean isChangedTree(List<Object> ownersTrace) {
-    if (changed && exist) {
-      return true;
-    }
-
-    ownersTrace.add(value);
-
-    if (value instanceof ODocument) {
-      ODocument doc = (ODocument) value;
-      return doc.isChangedInDepth();
-    }
-
-    if (value instanceof Collection) {
-      Collection list = (Collection) value;
-      for (Object element : list) {
-        if (element instanceof ODocument) {
-          ODocument doc = (ODocument) element;
-          for (Map.Entry<String, ODocumentEntry> field : doc._fields.entrySet()) {
-            if (field.getValue().isChangedTree(new ArrayList<>())) {
-              return true;
-            }
-          }
-        } else if (element instanceof Collection) {
-          if (ODocumentHelper.isChangedCollection((Collection) element, this, ownersTrace, 1)) {
-            return true;
-          }
-        } else if (element instanceof Map) {
-          if (ODocumentHelper.isChangedMap((Map<Object, Object>) element, this, ownersTrace, 1)) {
-            return true;
-          }
-        }
-      }
-    }
-
-    if (value instanceof Map) {
-      Map<Object, Object> map = (Map) value;
-      for (Map.Entry<Object, Object> entry : map.entrySet()) {
-        Object element = entry.getValue();
-        if (element instanceof ODocument) {
-          ODocument doc = (ODocument) element;
-          for (Map.Entry<String, ODocumentEntry> field : doc._fields.entrySet()) {
-            if (field.getValue().isChangedTree(new ArrayList<>())) {
-              return true;
-            }
-          }
-        } else if (element instanceof Collection) {
-          if (ODocumentHelper.isChangedCollection((List) element, this, ownersTrace, 1)) {
-            return true;
-          }
-        } else if (element instanceof Map) {
-          if (ODocumentHelper.isChangedMap((Map<Object, Object>) element, this, ownersTrace, 1)) {
-            return true;
-          }
-        }
-      }
-    }
-
-    if (timeLine != null) {
-      List<OMultiValueChangeEvent<Object, Object>> timeline = timeLine.getMultiValueChangeEvents();
-      if (timeline != null) {
-        for (OMultiValueChangeEvent<Object, Object> event : timeline) {
-          if (event.getChangeType() == OMultiValueChangeEvent.OChangeType.ADD
-              || event.getChangeType() == OMultiValueChangeEvent.OChangeType.NESTED
-              || event.getChangeType() == OMultiValueChangeEvent.OChangeType.UPDATE
-              || event.getChangeType() == OMultiValueChangeEvent.OChangeType.REMOVE) {
-            return true;
-          }
-        }
-      }
-    }
-
-    return false;
-  }
-
-  public boolean hasNonExistingTree() {
-    if (!exist) {
-      return true;
-    }
-
-    if (value instanceof ODocument) {
-      ODocument doc = (ODocument) value;
-      for (Map.Entry<String, ODocumentEntry> field : doc._fields.entrySet()) {
-        if (field.getValue().hasNonExistingTree()) {
-          return true;
-        }
-      }
-    }
-
-    if (value instanceof List) {
-      List list = (List) value;
-      for (Object element : list) {
-        if (element instanceof ODocument) {
-          ODocument doc = (ODocument) element;
-          for (Map.Entry<String, ODocumentEntry> field : doc._fields.entrySet()) {
-            if (field.getValue().hasNonExistingTree()) {
-              return true;
-            }
-          }
-        } else if (element instanceof List) {
-          if (ODocumentHelper.hasNonExistingInList((List) element)) {
-            return true;
-          }
-        }
-      }
-    }
-
-    return false;
   }
 
   public void setChanged(final boolean changed) {
