+++ core/src/main/java/com/orientechnologies/orient/core/record/impl/ODocumentHelper.java
@@ -1558,187 +1558,4 @@ public class ODocumentHelper {
     databaseRecord.activateOnCurrentThread();
     return function.call(databaseRecord);
   }
-
-  private static OMultiValueChangeEvent.OChangeType isNestedValueChanged(ONestedMultiValueChangeEvent event, Object value,
-      List<Object> ownersTrace, int ownersTraceOffset, Object valueIdentifier) {
-    if (event.getTimeLine() != null) {
-      List<OMultiValueChangeEvent<Object, Object>> events = event.getTimeLine().getMultiValueChangeEvents();
-      if (events != null) {
-        for (OMultiValueChangeEvent<Object, Object> nestedEvent : events) {
-          if (ownersTraceOffset < ownersTrace.size() && nestedEvent.getKey() == ownersTrace.get(ownersTraceOffset)
-              && nestedEvent instanceof ONestedMultiValueChangeEvent) {
-            ONestedMultiValueChangeEvent ne = (ONestedMultiValueChangeEvent) nestedEvent;
-            OMultiValueChangeEvent.OChangeType ret = isNestedValueChanged(ne, value, ownersTrace, ownersTraceOffset + 1,
-                valueIdentifier);
-            if (ret != null) {
-              return ret;
-            }
-          } else {
-            if (nestedEvent.getKey().equals(valueIdentifier) && nestedEvent.getValue() == value && ownersTraceOffset == ownersTrace
-                .size()) {
-              return nestedEvent.getChangeType();
-            }
-          }
-        }
-      }
-    }
-
-    return null;
-  }
-
-  public static OMultiValueChangeEvent.OChangeType isNestedValueChanged(ODocumentEntry entry, Object value,
-      List<Object> ownersTrace, int ownersTraceOffset, Object valueIdentifier) {
-    if (entry.timeLine != null) {
-      List<OMultiValueChangeEvent<Object, Object>> timeline = entry.timeLine.getMultiValueChangeEvents();
-      if (timeline != null) {
-        for (OMultiValueChangeEvent<Object, Object> event : timeline) {
-          if (ownersTraceOffset < ownersTrace.size() && event.getKey() == ownersTrace.get(ownersTraceOffset)
-              && event instanceof ONestedMultiValueChangeEvent) {
-            ONestedMultiValueChangeEvent nestedEvent = (ONestedMultiValueChangeEvent) event;
-            OMultiValueChangeEvent.OChangeType ret = isNestedValueChanged(nestedEvent, value, ownersTrace, ownersTraceOffset + 1,
-                valueIdentifier);
-            if (ret != null) {
-              return ret;
-            }
-          } else if ((event.getKey().equals(valueIdentifier) || event.getKey() == value) && event.getValue() == value
-              && ownersTraceOffset == ownersTrace.size()) {
-            return event.getChangeType();
-          }
-        }
-      }
-    }
-
-    return null;
-  }
-
-  private static boolean isInNestedEvent(ONestedMultiValueChangeEvent event, Object checkObject, List<Object> ownersTrace,
-      int ownersTraceOffset) {
-    if (event.getKey() == checkObject) {
-      return true;
-    }
-    if (event.getTimeLine() != null) {
-      List<OMultiValueChangeEvent<Object, Object>> timeline = event.getTimeLine().getMultiValueChangeEvents();
-      for (OMultiValueChangeEvent<Object, Object> nestedEvent : timeline) {
-        if (nestedEvent.getKey() == checkObject) {
-          return true;
-        } else {
-          if (nestedEvent instanceof ONestedMultiValueChangeEvent && isInNestedEvent((ONestedMultiValueChangeEvent) nestedEvent,
-              checkObject, ownersTrace, ownersTraceOffset + 1)) {
-            return true;
-          }
-        }
-      }
-    }
-    return false;
-  }
-
-  public static boolean isChangedCollection(Collection list, ODocumentEntry entry, List<Object> ownersTrace,
-      int ownersTraceOffset) {
-    ownersTrace.add(list);
-
-    //We can't use same logic as in isNestedValueChanged, because
-    //change of ODocument contained in list will not fire adequate change event
-    for (Object element : list) {
-      if (element instanceof ODocument) {
-        if (((ODocument) element).isChangedInDepth()) {
-          return true;
-        }
-      } else if (element instanceof Collection) {
-        if (isChangedCollection((Collection) element, entry, ownersTrace, ownersTraceOffset + 1)) {
-          return true;
-        }
-      } else if (element instanceof Map) {
-        if (isChangedMap((Map) element, entry, ownersTrace, ownersTraceOffset + 1)) {
-          return true;
-        }
-      }
-    }
-
-    if (entry.timeLine != null) {
-      List<OMultiValueChangeEvent<Object, Object>> timeline = entry.timeLine.getMultiValueChangeEvents();
-      if (timeline != null) {
-        for (OMultiValueChangeEvent<Object, Object> event : timeline) {
-          Object key = event.getKey();
-          if (key == list) {
-            return true;
-          }
-          if (event instanceof ONestedMultiValueChangeEvent && isInNestedEvent((ONestedMultiValueChangeEvent) event, list,
-              ownersTrace, ownersTraceOffset + 1)) {
-            return true;
-          }
-        }
-      }
-    }
-
-    return false;
-  }
-
-  public static boolean isChangedRidbag(ORidBag ridbag, ODocumentEntry entry) {
-
-    if (entry.timeLine != null) {
-      List<OMultiValueChangeEvent<Object, Object>> timeline = entry.timeLine.getMultiValueChangeEvents();
-      if (timeline != null) {
-        return !timeline.isEmpty();
-      }
-    }
-
-    return false;
-  }
-
-  public static boolean isChangedMap(Map<Object, Object> map, ODocumentEntry entry, List<Object> ownersTrace,
-      int ownersTraceOffset) {
-    ownersTrace.add(map);
-
-    //We can't use same logic as in isNestedValueChanged, because
-    //change of ODocument contained in list will not fire adequate change event
-    for (Map.Entry<Object, Object> mapEntry : map.entrySet()) {
-      Object element = mapEntry.getValue();
-      if (element instanceof ODocument) {
-        if (((ODocument) element).isChangedInDepth()) {
-          return true;
-        }
-      } else if (element instanceof Collection) {
-        if (isChangedCollection((Collection) element, entry, ownersTrace, ownersTraceOffset + 1)) {
-          return true;
-        }
-      } else if (element instanceof Map) {
-        if (isChangedMap((Map) element, entry, ownersTrace, ownersTraceOffset + 1)) {
-          return true;
-        }
-      }
-    }
-
-    if (entry.timeLine != null) {
-      List<OMultiValueChangeEvent<Object, Object>> timeline = entry.timeLine.getMultiValueChangeEvents();
-      if (timeline != null) {
-        for (OMultiValueChangeEvent<Object, Object> event : timeline) {
-          Object key = event.getKey();
-          if (key == map) {
-            return true;
-          }
-          if (event instanceof ONestedMultiValueChangeEvent && isInNestedEvent((ONestedMultiValueChangeEvent) event, map,
-              ownersTrace, ownersTraceOffset + 1)) {
-            return true;
-          }
-        }
-      }
-    }
-
-    return false;
-  }
-
-  protected static boolean hasNonExistingInList(List list) {
-    for (Object element : list) {
-      if (element instanceof ODocument) {
-        if (((ODocument) element).hasNonExistingInDepth()) {
-          return true;
-        }
-      } else if (element instanceof List) {
-        if (hasNonExistingInList((List) element)) {
-          return true;
-        }
-      }
-    }
-    return false;
-  }
 }
