+++ core/src/main/java/com/orientechnologies/orient/core/serialization/serializer/record/binary/ORecordSerializerNetworkV37.java
@@ -31,7 +31,6 @@ import com.orientechnologies.orient.core.config.OGlobalConfiguration;
 import com.orientechnologies.orient.core.db.ODatabaseRecordThreadLocal;
 import com.orientechnologies.orient.core.db.record.*;
 import com.orientechnologies.orient.core.db.record.ridbag.ORidBag;
-import com.orientechnologies.orient.core.delta.ODocumentDelta;
 import com.orientechnologies.orient.core.exception.OSerializationException;
 import com.orientechnologies.orient.core.exception.OValidationException;
 import com.orientechnologies.orient.core.id.ORID;
@@ -867,14 +866,8 @@ public class ORecordSerializerNetworkV37 implements ORecordSerializer {
     } else {
       final BytesContainer container = new BytesContainer();
 
-      ODocument doc = (ODocument) iSource;
       // SERIALIZE RECORD
-      if (!iOnlyDelta) {
-        serialize(doc, container, false);
-      } else {
-        ODocumentDelta deltaDoc = doc.getDeltaFromOriginal();
-        return deltaDoc.serialize();
-      }
+      serialize((ODocument) iSource, container, false);
 
       return container.fitBytes();
     }
