+++ distributed/src/main/java/com/orientechnologies/orient/server/distributed/impl/ODatabaseDocumentDistributed.java
@@ -741,19 +741,11 @@ public class ODatabaseDocumentDistributed extends ODatabaseDocumentEmbedded {
     commit2pc(transactionId);
   }
 
-  /**
-   * @param transactionId
-   *
-   * @return null returned means that commit failed
-   */
-  public Collection<ORecordOperation> commit2pc(ODistributedRequestId transactionId) {
+  public boolean commit2pc(ODistributedRequestId transactionId) {
     getStorageDistributed().resetLastValidBackup();
     ODistributedDatabase localDistributedDatabase = getStorageDistributed().getLocalDistributedDatabase();
-
     ODistributedServerManager manager = getStorageDistributed().getDistributedManager();
     ONewDistributedTxContextImpl txContext = (ONewDistributedTxContextImpl) localDistributedDatabase.getTxContext(transactionId);
-    Collection<ORecordOperation> operations = txContext.getTransaction().getRecordOperations();
-
     if (txContext != null) {
       if (SUCCESS.equals(txContext.getStatus())) {
         try {
@@ -765,7 +757,7 @@ public class ODatabaseDocumentDistributed extends ODatabaseDocumentEmbedded {
           OLiveQueryHook.removePendingDatabaseOps(this);
           OLiveQueryHookV2.removePendingDatabaseOps(this);
         }
-        return operations;
+        return true;
       } else if (TIMEDOUT.equals(txContext.getStatus())) {
         for (int i = 0; i < 10; i++) {
           try {
@@ -804,7 +796,7 @@ public class ODatabaseDocumentDistributed extends ODatabaseDocumentEmbedded {
         });
       }
     }
-    return null;
+    return false;
   }
 
   public boolean rollback2pc(ODistributedRequestId transactionId) {
