+++ distributed/src/main/java/com/orientechnologies/orient/server/distributed/impl/task/OTransactionPhase2Task.java
@@ -2,17 +2,8 @@ package com.orientechnologies.orient.server.distributed.impl.task;
 
 import com.orientechnologies.common.log.OLogManager;
 import com.orientechnologies.orient.core.Orient;
-import com.orientechnologies.orient.core.cache.OLocalRecordCache;
 import com.orientechnologies.orient.core.command.OCommandDistributedReplicateRequest;
 import com.orientechnologies.orient.core.db.ODatabaseDocumentInternal;
-import com.orientechnologies.orient.core.db.ODatabaseSession;
-import com.orientechnologies.orient.core.db.record.OIdentifiable;
-import com.orientechnologies.orient.core.db.record.ORecordOperation;
-import com.orientechnologies.orient.core.delta.ODocumentDelta;
-import com.orientechnologies.orient.core.id.ORID;
-import com.orientechnologies.orient.core.record.ORecord;
-import com.orientechnologies.orient.core.record.ORecordAbstract;
-import com.orientechnologies.orient.core.record.ORecordInternal;
 import com.orientechnologies.orient.core.storage.impl.local.paginated.wal.OLogSequenceNumber;
 import com.orientechnologies.orient.server.OServer;
 import com.orientechnologies.orient.server.distributed.ODistributedRequestId;
@@ -24,7 +15,6 @@ import com.orientechnologies.orient.server.distributed.task.OAbstractReplicatedT
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
-import java.util.Collection;
 
 import static com.orientechnologies.orient.core.config.OGlobalConfiguration.DISTRIBUTED_CONCURRENT_TX_MAX_AUTORETRY;
 
@@ -95,38 +85,11 @@ public class OTransactionPhase2Task extends OAbstractReplicatedTask {
     }
   }
 
-  private void updateRecordVersionsinCache(ODatabaseDocumentInternal database, Collection<ORecordOperation> operations) {
-
-    if (database instanceof ODatabaseSession) {
-      ODatabaseSession session = (ODatabaseSession) database;
-      OLocalRecordCache cache = session.getLocalCache();
-      for (ORecordOperation operation : operations) {
-        Object resulData = operation.getResultData();
-        if (resulData instanceof Integer) {
-          Integer version = (Integer) resulData;
-          if (version != null) {
-            OIdentifiable operationRecord = operation.getRecordContainer();
-            if (operationRecord instanceof ODocumentDelta) {
-              ODocumentDelta deltaRecord = (ODocumentDelta) operationRecord;
-              ORID id = deltaRecord.getIdentity();
-              ORecord rec = cache.findRecord(id);
-              if (rec != null && rec instanceof ORecordAbstract) {
-                ORecordInternal.setVersion(rec, version);
-              }
-            }
-          }
-        }
-      }
-    }
-
-  }
-
   @Override
   public Object execute(ODistributedRequestId requestId, OServer iServer, ODistributedServerManager iManager,
       ODatabaseDocumentInternal database) throws Exception {
     if (success) {
-      Collection<ORecordOperation> operations = ((ODatabaseDocumentDistributed) database).commit2pc(transactionId);
-      if (operations == null) {
+      if (!((ODatabaseDocumentDistributed) database).commit2pc(transactionId)) {
         retryCount++;
         if (retryCount < database.getConfiguration().getValueAsInteger(DISTRIBUTED_CONCURRENT_TX_MAX_AUTORETRY)) {
           OLogManager.instance()
@@ -146,9 +109,6 @@ public class OTransactionPhase2Task extends OAbstractReplicatedTask {
         }
       } else {
         hasResponse = true;
-        if (OTransactionPhase1Task.useDeltasForUpdate) {
-          updateRecordVersionsinCache(database, operations);
-        }
       }
     } else {
       if (!((ODatabaseDocumentDistributed) database).rollback2pc(transactionId)) {
