+++ distributed/src/test/java/com/orientechnologies/orient/server/distributed/impl/OTransactionPhase1TaskTest.java
@@ -93,7 +93,6 @@ public class OTransactionPhase1TaskTest {
   public void testExecutionConcurrentModificationUpdate() throws Exception {
     ODocument doc = new ODocument("TestClass");
     doc.field("first", "one");
-    doc.field("second", "constantFieldValue");
     session.save(doc);
     ODocument old = doc.copy();
     doc.field("first", "two");
